+++ app/css/index.css
@@ -113,7 +113,6 @@ body {
 	box-shadow: 0 0 20px #3cb371;
 }
 
-
 .paragraph {
 	margin: 0px;
 }
