--- app/ts/ens.ts
@@ -1,8 +1,0 @@
-export const MOCK_ADDRESS = '0xdeadbeefdeadbeefdeadbeefdeadbeefdeadbeef'
-export const ENS_PUBLIC_RESOLVER = '0x231b0Ee14048e9dCcD1d247744d114a4EB5E8E63'
-export const ENS_TOKEN_WRAPPER = '0xD4416b13d2b3a9aBae7AcD5D6C2BbDBE25686401'
-export const ENS_ETH_REGISTRAR_CONTROLLER = '0x253553366Da8546fC250F225fe3d25d0C782303b'
-export const ENS_ETHEREUM_NAME_SERVICE = '0x57f1887a8BF19b14fC0dF6Fd9B2acc9Af147eA85'
-export const ENS_PUBLIC_RESOLVER_2 = '0x4976fb03C32e5B8cfe2b6cCB31c09Ba78EBaBa41'
-export const ENS_REGISTRY_WITH_FALLBACK = '0x00000000000C2E074eC69A0dFb2997BA6C7d2e1e'
-export const ENS_REVERSE_REGISTRAR = '0xa58E81fe9b61B5c3fE2AFD33CF304c454AbFc7Cb'
